eval: Share environment when creating Function objects
This fixes a bug that prevented a function object from referencing
itself (i.e. recurrsion in function objects). The bug was due to the
fact we make a copy of the environment, and at that point the function
is not defined, so the environment the function uses does not contain
itself.